+++ docs/docusaurus.config.js
@@ -1,0 +1,103 @@
+// @ts-check
+// Note: type annotations allow type checking and IDEs autocompletion
+
+const lightCodeTheme = require('prism-react-renderer/themes/github');
+const darkCodeTheme = require('prism-react-renderer/themes/dracula');
+
+/** @type {import('@docusaurus/types').Config} */
+const config = {
+  title: 'Soql Parser JS',
+  tagline: 'Parse and generate SOQL queries',
+  url: 'https://paustint.github.io',
+  baseUrl: '/soql-parser-js/',
+  onBrokenLinks: 'throw',
+  onBrokenMarkdownLinks: 'warn',
+  favicon: 'img/favicon.ico',
+
+  // GitHub pages deployment config.
+  // If you aren't using GitHub pages, you don't need these.
+  organizationName: 'paustint',
+  projectName: 'soql-parser-js',
+  deploymentBranch: 'gh-pages',
+  trailingSlash: false,
+
+  // Even if you don't use internalization, you can use this field to set useful
+  // metadata like html lang. For example, if your site is Chinese, you may want
+  // to replace "en" with "zh-Hans".
+  i18n: {
+    defaultLocale: 'en',
+    locales: ['en'],
+  },
+
+  presets: [
+    [
+      'classic',
+      /** @type {import('@docusaurus/preset-classic').Options} */
+      ({
+        docs: {
+          sidebarPath: require.resolve('./sidebars.js'),
+          // Please change this to your repo.
+          // Remove this to remove the "edit this page" links.
+          editUrl: 'https://github.com/paustint/soql-parser-js/docs/docs/',
+        },
+        theme: {
+          customCss: require.resolve('./src/css/custom.css'),
+        },
+      }),
+    ],
+  ],
+
+  themeConfig:
+    /** @type {import('@docusaurus/preset-classic').ThemeConfig} */
+    ({
+      colorMode: {
+        defaultMode: 'dark',
+        respectPrefersColorScheme: true,
+      },
+      navbar: {
+        title: 'Home',
+        logo: {
+          alt: 'Logo',
+          src: 'img/soql-parser-js-logo.svg',
+        },
+        items: [
+          {
+            type: 'doc',
+            docId: 'overview',
+            position: 'left',
+            label: 'Documentation',
+          },
+          { to: '/playground', label: 'Playground', position: 'left' },
+          {
+            href: 'https://github.com/paustint/soql-parser-js',
+            label: 'GitHub',
+            position: 'right',
+          },
+        ],
+      },
+      footer: {
+        style: 'dark',
+        links: [
+          {
+            title: 'Links',
+            items: [
+              {
+                label: 'Documentation',
+                to: '/docs/overview',
+              },
+              {
+                label: 'GitHub',
+                href: 'https://github.com/paustint/soql-parser-js',
+              },
+            ],
+          },
+        ],
+      },
+      prism: {
+        theme: lightCodeTheme,
+        darkTheme: darkCodeTheme,
+      },
+    }),
+};
+
+module.exports = config;