+++ docs/src/types.ts
@@ -1,5 +1,0 @@
-export interface Query {
-    key: string;
-    active: boolean;
-    query: string;
-}